+++ tool/build_actions.dart
@@ -23,6 +23,7 @@ final List<BuildAction> buildActions = [
         'lib/src/analyzer_output.dart',
         'lib/src/fitness.dart',
         'lib/src/license.dart',
+        'lib/src/maintenance.dart',
         'lib/src/platform.dart',
         'lib/src/pub_summary.dart',
         'lib/src/summary.dart',